Sent video does not contains duration (#3272) When using the file picker (and not the media picker). Now when using the file picker, we detect the mime type and we send the correct event Also some code duplication
